+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
   **For upgrading:** the easiest way to fix the build is to remove the offending `@ref` links. Alternatively, the `repo` argument to `makedocs` can be set to the appropriate `Remotes.Remote` object that implements the `Remotes.issueurl` function, which would make sure that correct URLs are generated.
 
+* ![Enhancement][badge-enhancement] Woodpecker CI is now automatically supported for documentation deployment. ([#1880][github-1880])
 * ![Bugfix][badge-bugfix] Documenter now generates the correct source URLs for docstrings from other packages when the `repo` argument to `makedocs` is set (note: the source links to such docstrings only work if the external package is cloned from GitHub and added as a dev-dependency). However, this change **breaks** the case where the `repo` argument is used to override the main package/repository URL, assuming the repository is cloned from GitHub. ([#1808][github-1808])
 * ![Bugfix][badge-bugfix] Documenter no longer uses the `TRAVIS_REPO_SLUG` environment variable to determine the Git remote of non-main repositories (when inferring it from the Git repository configuration has failed), which could previously lead to bad source links. ([#1881][github-1881])
 
@@ -1105,6 +1106,7 @@
 [github-1865]: https://github.com/JuliaDocs/Documenter.jl/pull/1865
 [github-1870]: https://github.com/JuliaDocs/Documenter.jl/issues/1870
 [github-1871]: https://github.com/JuliaDocs/Documenter.jl/pull/1871
+[github-1880]: https://github.com/JuliaDocs/Documenter.jl/pull/1880
 [github-1881]: https://github.com/JuliaDocs/Documenter.jl/pull/1881
 <!-- end of issue link definitions -->